Improved & Fixed models classes of Orders & Cart & Address + Fixed header profile link & home link + Orders are now created on cart confirmation + An address is asked on cart confirmation. Ref #17 #42 #46

+++ mysql-dump/init.sql
@@ -112,9 +112,10 @@ CREATE TABLE IF NOT EXISTS Customer (
     FOREIGN KEY (idLoyaltyAccount) REFERENCES LoyaltyAccount (id)
 );
 
+
 CREATE TABLE IF NOT EXISTS Orders (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    total INT NOT NULL,
+    total FLOAT NOT NULL,
     date DATE NOT NULL,
     orderStatus VARCHAR(30) NOT NULL,
     idCustomer INT NOT NULL,
